feat: Pass all dashBoard in Monopage, creation page gestion agenda
