--- lib/screens/connexion_validate.dart
@@ -1,12 +1,0 @@
-import 'package:flutter/material.dart';
-
-class ConnexionValidate extends StatelessWidget {
-  const ConnexionValidate({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return const Scaffold(
-      body: Center(child: Text("ConnexionValidate")),
-    );
-  }
-}
