+++ pubspec.yaml
@@ -8,23 +8,22 @@
 # following page: https://dart.dev/tools/pub/pubspec
 
 dependencies:
+  intl: ^0.17.0
   cupertino_icons: ^1.0.2
   flutter_dotenv: ^5.0.2
   shared_preferences: ^2.0.8
-  simple_gradient_text: 1.2.4
+  simple_gradient_text: ^1.3.0
   dart_jsonwebtoken: ^2.11.0
-  chat_bubbles: ^1.3.1
   font_awesome_flutter: ^10.3.0
-  flutter_svg: ^1.1.6
+  flutter_svg: ^2.0.7
   graphql: ^5.0.0
-  google_maps_flutter: ^2.1.1
   http: ^1.1.0
   path_provider: ^2.0.5
   flutter_pdfview: ^1.0.0
   flutter:
     sdk: flutter
   flutter_launcher_icons: ^0.13.1
-  flutter_local_notifications: ^14.0.1
+  flutter_local_notifications: ^16.1.0
 description: A new Flutter project.
 dev_dependencies:
   flutter_lints: ^2.0.0
