+++ test/annuaire_medecin_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:prototype_1/screens/annuaire_medecin.dart';
+import 'package:prototype_1/screens/LandingPage/annuaire_medecin.dart';
 
 void main() {
   testWidgets(
