+++ test/help_page_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:prototype_1/screens/help.dart';
+import 'package:prototype_1/screens/LandingPage/help.dart';
 
 void main() {
   testWidgets('Test fonctionnel pour HelpScreen', (WidgetTester tester) async {
