+++ test/landing_page_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:prototype_1/screens/landing_page.dart';
+import 'package:prototype_1/screens/LandingPage/landing_page.dart';
 
 void main() {
   testWidgets('Test fonctionnel de la page d\'accueil',
